pyproject.toml
[tool.poetry]
name = "spacex-test"
version = "0.1.0"
description = ""
authors = ["Aimene AHDIBI <aimene.ahdibi@orange.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.8.1"
flask-restx = "^1.1.0"
peewee = "^3.16.0"
python-dotenv = "^1.0.0"
flask = "^2.2.3"
pytest = "^7.2.2"
psycopg2-binary = "^2.9.6"
beautifulsoup4 = "^4.12.2"
requests = "^2.28.2"
uwsgi = "^2.0.21"


[tool.poetry.group.dev.dependencies]
pyproject-toml = "^0.0.10"
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
